Disallows root_fs updates
While it's useful to provide an initial config for root_fs, there are a
couple of concerns with passing it to the update api. I've found a few
strange behaviours after updating the root fs size. One of the most
concerning is that while the UI shows the correctly resized value
following an update, when subsequently using the UI resize
functionality, the new size appears to be relative to the old size (e.g.
8GB -> 12GB through API, resize +3GB in UI -> 11GB instead of the
expected 15GB).

It may be worth exploring using the separate resize API to handle this
"properly", but that complicates things and still doesn't handle the
other issue in that there is a mismatch with the initially configured
root_fs, and the value that eventually gets configured (e.g. local-lvm:8
becomes something like local-lvm:vm-100-disk-1,size=8G). When creating a
new container with the "fully qualified" root_fs, it complains about the
disk not existing during creation. There's obviously some extra step
happening after inital creation which is reformatting the root_fs in the
example above.

Considering file system resizing is something that probably won't happen
as often as the other updates, and currently tends to involve a lot more
steps than just changing a value in a simple UI, this seems like a low
priority and can be ignored for now.

I have added an error when the root_fs is changed that halts the update.