Cleanup resourceContainer and handles start/stop
Cleanup formatting in resourceContainer, and adds start/stop to create
and destroy. After creation, a container will start. Before deletion, we
check the container's status, and shut it down if necessary.

Startup is a convenience (potentially configurable in the future), while
shutdown is a necessity. If you try to delete a container that is
running, the API returns an error which halts the terraform apply.