Adds proxmox_backup
Due to API limitations, only create is currently implemented. Basically,
for resources with auto-generated ids (backups being one of them), it's
impossible to figure out the id of the resource you just created. This
is because all non-GET methods seem to return null.

I have posted this question on the forums, along with a few others:

https://forum.proxmox.com/threads/api-questions.37581/